Move Audit class into Audits module
This is a first step to moving classes into an Audits namespace.
In the future we will iterate this further as right now we end up
with a class named Audits::Audit from this commit,
which is not ideal.
