+++ app/models/content_item.rb
@@ -1,5 +1,5 @@
 class ContentItem < ApplicationRecord
-  has_one :audit, primary_key: :content_id, foreign_key: :content_id
+  has_one :audit, primary_key: :content_id, foreign_key: :content_id, class_name: "Audits::Audit"
   has_one :report_row, primary_key: :content_id, foreign_key: :content_id
   has_many :links, primary_key: :content_id, foreign_key: :source_content_id
