+++ app/views/audits/audits/_boolean_question.html.erb
