+++ app/views/audits/audits/_content_item.html.erb
