+++ app/views/audits/audits/_progress_bar.html.erb
