+++ app/views/audits/audits/_sidebar.html.erb
